pyproject.toml
[tool.ruff]
src = ["src"]
line-length = 120

extend-exclude = ["!**", ".history", "versions"]

# flake8 previously used the following codes:
# E701 E231 E225 E999 W503 E251 C901 I004 E800 B008 B024 B026 B028
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming (N)
    "YTT", # flake8-2020
    "C",   # flake8-comprehensions
    "B",   # flake8-bugbear
    "T20", # flake8-print
    "RET", # flake8-return
    "SLF", # flake8-self
    "ARG", # flake8-unused-arguments
    "TCH", # flake8-type-checking
    "RUF", # Ruff-specific rules
]

ignore = [
    "B008", # do not perform function calls in argument defaults
    "C901", # too complex
    "B904", # `except` clause, raise exceptions with `raise,
    "B027", # is an empty method in an abstract base class
    "B024", # abstract base class, but it has no abstract methods
    "B026", # Star-arg unpacking after a keyword argument is strongly discouraged
    "B905", # `zip()` without an explicit `strict=` parameter
    "E701", # multiple statements on one line
]

[tool.ruff.lint.mccabe]
# Unlike Flake8, default to a complexity level of 10.
max-complexity = 10

# ISORT CONFIGURATION USED BY RUFF
[tool.ruff.lint.isort] # https://beta.ruff.rs/docs/settings/#isort
force-sort-within-sections = true

force-wrap-aliases = true
combine-as-imports = true

# The sections and order of the imports
section-order = [
    "future",
    "standard-library",
    "third-party",
    "first-party",
    "local-folder",
]

# Our libraries used on the project, must be added here, then they will be groupped together
known-first-party = []

# Our local folders used on the project, must be added here, then they will be groupped together
known-local-folder = [
    "app",
    "auth",
    "config",
    "core",
    "exam_guard",
    "infra",
    "shared",
    "utils",
]

# Third party libraries used on the project, must be added here, then they will be groupped together
known-third-party = [
    "beartype",
    "bcrypt",
    "colorlog",
    "fastapi",
    "pydantic",
    "pytest",
    "typer",
    "yaml",
]

lines-after-imports = 2
# lines-between-types = 1

[tool.ruff.lint.flake8-bugbear]
extend-immutable-calls = ["fastapi.Depends", "fastapi.Query"]


[tool.ruff.format]
quote-style = "single"
indent-style = "space"
docstring-code-format = true

[tool.black]
line-length = 120
skip-string-normalization = true
target-version = ['py312']
include = '\.pyi?$'
exclude = '''

(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
    | migrations
  )/
  | foo.py           # also separately exclude a file named foo.py in
                     # the root of the project
)
'''

[tool.pylint]
max-line-length = 120
disable = ["C0112", "C0114", "C0115", "C0116"]


[tool.poetry]
name = "exam-guard"
version = "0.1.0"
description = ""
authors = ["Alexandre Busquets Triola <abusquets@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.110.1"
uvicorn = "^0.29.0"
pydantic = "^2.6.4"
pyyaml = "^6.0"
colorlog = "^6.7.0"
sqlalchemy = "^2.0.29"
alembic = "^1.13.1"
bcrypt = "^4.0.1"
asyncpg = "0.29"
httpx = "0.26.0"
pydantic-settings = "^2.1.0"
greenlet = "^3.0.3"
faker = "22.0.0"
anyio = "^4.2.0"
pytest = "^7.4.4"
dpath = "^2.1.6"

[tool.poetry.group.dev.dependencies]
debugpy = "^1.8.1"
pytest = "^7.4.3"
pytest-env = "^1.1.3"
pytest-cov = "^4.1.0"
pytest-asyncio = "^0.23.2"
poetry-types = "^0.5.0"
types-redis = "^4.6.0.11"
watchdog = "^4.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
